pyproject.toml
[project]
name = "cnn-gru-seqcap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.5",
    "nvitop>=1.5.3",
    "torch==2.3.1",
    "torchaudio==2.3.1",
    "torchvision==0.18.1",
    "tqdm>=4.67.1",
    "pillow>=8.0.0",
    "pyyaml>=6.0.2",
    "pandas>=2.3.2",
    "librosa>=0.11.0",
    "pysndfx>=0.3.6",
    "transformers>=4.55.4",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
